requirements.txt
flask==2.0.1
tensorflow==2.2.0
numpy==1.18.1
pillow==7.1.1
keras==2.3.1
sklearn